ios fix pt6
