+++ src/test/java/com/solvd/tests/demoapp/ios/MyDemoAppIOSTest.java
@@ -1,0 +1,25 @@
+package com.solvd.tests.demoapp.ios;
+
+import com.solvd.demoapp.pages.common.CatalogPageBase;
+import com.solvd.demoapp.pages.common.LoginPageBase;
+import com.solvd.demoapp.pages.common.MenuPageBase;
+import com.solvd.tests.AbstractTest;
+import org.testng.annotations.Test;
+
+import static org.testng.Assert.assertTrue;
+
+public class MyDemoAppIOSTest extends AbstractTest {
+
+    @Test(testName = "ANDT-13", description = "Validate that user can log in")
+    public void validateLogging() {
+        CatalogPageBase catalogPage = initPage(getDriver(), CatalogPageBase.class);
+        MenuPageBase MenuPage = catalogPage.clickMoreButton();
+        LoginPageBase loginPage = MenuPage.goToLoginPage();
+        CatalogPageBase catalogPageLogged = loginPage.loginBob();
+        boolean pageOpened = catalogPageLogged.isPageOpened();
+        assertTrue(pageOpened, "Successfully redirected after logging in");
+        MenuPageBase rightMenuPageLogged = catalogPageLogged.clickMoreButton();
+        boolean visible = rightMenuPageLogged.getLogoutButton().isVisible(0);
+        assertTrue(visible, "Logout button is not displayed - user not logged!");
+    }
+}
